tinder42/requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
python-dotenv
email-validator
Flask-Mail
